Fixed date placeholder and improved roles check
